# worldwide-framework/openapi-worldwide/src/main/resources/openapi/json/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "世界的领域模型",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/mall/product/category/{category_id}": {
      "get": {
        "summary": "获取商品分类内容",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "获取商品分类内容",
        "operationId": "getProductCategoryContent",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "description": "商品分类编码",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductCategoryResp"
                },
                "examples": {}
              }
            }
          }
        }
      },
      "put": {
        "summary": "更新商品分类",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "更新商品分类",
        "operationId": "updateProductCategory",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [
          {
            "name": "category_id",
            "in": "path",
            "description": "商品分类id",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductCategoryReq",
                "title": "更新商品分类DTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
...
    "/mall/product/brand/{brand_id}": {
      "put": {
        "summary": "更新商品品牌",
        "x-apifox-folder": "mall/product/brand",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "更新商品品牌",
        "operationId": "updateProductBrand",
        "tags": [
          "mall/product/brand"
        ],
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "description": "商品品牌编码",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductBrandReq"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                },
                "examples": {}
              }
            }
          }
        }
      },
      "get": {
        "summary": "获取商品品牌内容",
        "x-apifox-folder": "mall/product/brand",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "获取商品品牌内容",
        "operationId": "getProductBrandContent",
        "tags": [
          "mall/product/brand"
        ],
        "parameters": [
          {
            "name": "brand_id",
            "in": "path",
            "description": "商品品牌编码",
            "required": true,
            "example": "1",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductBrandResp"
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/mall/product/category": {
      "post": {
        "summary": "添加商品分类",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "添加商品分类",
        "operationId": "createProductCategory",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductCategoryReq",
                "title": "创建商品分类请求DTO",
                "x-apifox-overrides": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/account/admin/{sys_admin_id}": {
      "post": {
        "summary": "超级管理员添加系统管理员",
        "x-apifox-folder": "account/admin",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "超级管理员添加系统管理员",
        "operationId": "SysAdminAddAdminUser",
        "tags": [
          "account/admin"
        ],
        "parameters": [
          {
            "name": "sys_admin_id",
            "in": "path",
            "description": "超级管理员编码",
            "required": true,
            "example": "",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddAdminReq",
                "x-apifox-overrides": {}
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/mall/product/category/page": {
      "get": {
        "summary": "分页获取商品分类内容",
        "x-apifox-folder": "mall/product/category",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "分页获取商品分类内容",
        "operationId": "pageProductCategoryContent",
        "tags": [
          "mall/product/category"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageQueryProductCategoryReq",
                "x-apifox-overrides": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductCategoryResp",
                  "x-apifox-overrides": {}
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
...
    "/mall/product/brand/page": {
      "get": {
        "summary": "分页查询商品品牌",
        "x-apifox-folder": "mall/product/brand",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "分页查询商品品牌",
        "operationId": "pageProductBrandContent",
        "tags": [
          "mall/product/brand"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PageQueryProductBrandReq",
                "x-apifox-overrides": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PageProductBrandResp",
                  "x-apifox-overrides": {}
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
    "/mall/product/brand": {
      "post": {
        "summary": "添加商品品牌",
        "x-apifox-folder": "mall/product/brand",
        "x-apifox-status": "developing",
        "deprecated": false,
        "description": "添加商品品牌",
        "operationId": "createProductBrand",
        "tags": [
          "mall/product/brand"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProductBrandReq",
                "x-apifox-overrides": {},
                "title": "创建商品品牌DTO"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "成功",
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer"
                },
                "examples": {}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PageProductCategoryResp": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "title": "条数"
          },
          "data": {
            "title": "商品分类内容集",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductCategoryResp"
            }
          }
        },
        "x-apifox-orders": [
          "total",
          "data"
        ],
        "required": [
          "total",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
      "ProductCategoryResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "title": "商品分类编码"
          },
          "pid": {
            "type": "integer",
            "title": "商品分类父编码"
          },
          "name": {
            "type": "string",
            "title": "商品分类名称"
          },
          "description": {
            "type": "string",
            "title": "商品分类描述"
          },
          "picUrl": {
            "type": "string",
            "title": "商品分类图片"
          },
          "sort": {
            "type": "integer",
            "title": "商品分类排序"
          },
          "status": {
            "title": "状态",
            "description": "0->关闭;1->开启;",
            "$ref": "#/components/schemas/StatusEnum"
          },
          "deleted": {
            "title": "是否被删除",
            "description": "0->已删除;1->未删除;",
            "$ref": "#/components/schemas/DeletedEnum"
          },
          "createTime": {
            "type": "string",
            "title": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "title": "更新时间"
          }
        },
        "required": [
          "id",
          "pid",
          "name",
          "sort",
          "status",
          "deleted",
          "createTime"
        ],
        "x-apifox-orders": [
          "id",
          "pid",
          "name",
          "description",
          "picUrl",
          "sort",
          "status",
          "deleted",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
...
      "ProductBrandResp": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "title": "品牌编码"
          },
          "name": {
            "type": "string",
            "title": "品牌名称"
          },
          "description": {
            "type": "string",
            "title": "品牌描述"
          },
          "picUrl": {
            "type": "string",
            "title": "品牌图片"
          },
          "status": {
            "title": "状态",
            "description": "0->关闭;1->开启;",
            "$ref": "#/components/schemas/StatusEnum"
          },
          "deleted": {
            "title": "是否被删除",
            "description": "0->已删除;1->未删除;",
            "$ref": "#/components/schemas/DeletedEnum"
          },
          "createTime": {
            "type": "string",
            "title": "创建时间"
          },
          "updateTime": {
            "type": "string",
            "title": "更新时间"
          }
        },
        "required": [
          "id",
          "name",
          "status",
          "deleted",
          "createTime"
        ],
        "x-apifox-orders": [
          "id",
          "name",
          "description",
          "picUrl",
          "status",
          "deleted",
          "createTime",
          "updateTime"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/brand"
      },
      "CreateProductCategoryReq": {
        "type": "object",
        "properties": {
          "pid": {
            "type": "integer",
            "title": "父分类编码"
          },
          "name": {
            "type": "string",
            "title": "分类名称",
            "minLength": 1,
            "maxLength": 20
          },
          "description": {
            "type": "string",
            "title": "分类描述"
          },
          "picUrl": {
            "type": "string",
            "title": "分类图片"
          },
          "sort": {
            "type": "integer",
            "title": "分类排序",
            "minimum": 0
          }
        },
        "required": [
          "pid",
          "name",
          "sort"
        ],
        "x-apifox-orders": [
          "pid",
          "name",
          "description",
          "picUrl",
          "sort"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
      "AddAdminReq": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "title": "用户账号"
          },
          "password": {
            "type": "string",
            "title": "用户密码"
          }
        },
        "x-apifox-orders": [
          "username",
          "password"
        ],
        "required": [
          "username",
          "password"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "account/admin"
      },
...
      "PageQueryProductBrandReq": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "title": "页码",
            "minimum": 1
          },
          "pageSize": {
            "type": "integer",
            "title": "页面大小",
            "minimum": 1
          },
          "id": {
            "type": "integer",
            "title": "品牌编码"
          },
          "name": {
            "type": "string",
            "title": "品牌名称"
          },
          "status": {
            "title": "状态",
            "description": "0->关闭;1->开启;",
            "$ref": "#/components/schemas/StatusEnum"
          }
        },
        "required": [
          "pageNo",
          "pageSize"
        ],
        "x-apifox-orders": [
          "pageNo",
          "pageSize",
          "id",
          "name",
          "status"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/brand"
      },
...
      "PageQueryProductCategoryReq": {
        "type": "object",
        "properties": {
          "pageNo": {
            "type": "integer",
            "title": "页码"
          },
          "pageSize": {
            "type": "integer",
            "title": "页面大小"
          },
          "id": {
            "type": "integer",
            "title": "分类编码"
          },
          "pid": {
            "type": "integer",
            "title": "分类父编码"
          },
          "name": {
            "type": "string",
            "title": "分类名称"
          },
          "status": {
            "title": "状态",
            "description": "0->关闭;1->开启;",
            "$ref": "#/components/schemas/StatusEnum"
          }
        },
        "x-apifox-orders": [
          "pageNo",
          "pageSize",
          "id",
          "pid",
          "name",
          "status"
        ],
        "required": [
          "pageNo",
          "pageSize"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
      "PageProductBrandResp": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "title": "条数"
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductBrandResp"
            },
            "title": "商品品牌内容数据集"
          }
        },
        "required": [
          "total",
          "data"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-orders": [
          "total",
          "data"
        ],
        "x-apifox-folder": "mall/product/brand"
      },
      "UpdateProductCategoryReq": {
        "type": "object",
        "properties": {
          "pid": {
            "type": "integer",
            "title": "父分类编码"
          },
          "name": {
            "type": "string",
            "title": "分类名称",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "title": "分类描述"
          },
          "picUrl": {
            "type": "string",
            "title": "分类图片"
          },
          "sort": {
            "type": "integer",
            "title": "分类排序"
          },
          "status": {
            "title": "状态",
            "description": "0->关闭;1->开启;",
            "$ref": "#/components/schemas/StatusEnum"
          }
        },
        "x-apifox-orders": [
          "pid",
          "name",
          "description",
          "picUrl",
          "sort",
          "status"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/category"
      },
...
      "StatusEnum": {
        "type": "integer",
        "enum": [
          0,
          1
        ],
        "x-apifox": {
          "enumDescriptions": {
            "0": "关闭",
            "1": "开启"
          }
        },
        "x-apifox-folder": "enums"
      },
...
      "CreateProductBrandReq": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "title": "品牌名称",
            "minLength": 1,
            "maxLength": 20
          },
          "description": {
            "type": "string",
            "title": "品牌描述"
          },
          "picUrl": {
            "type": "string",
            "title": "品牌名图片"
          }
        },
        "x-apifox-orders": [
          "name",
          "description",
          "picUrl"
        ],
        "required": [
          "name"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/brand"
      },
...
      "DeletedEnum": {
        "type": "integer",
        "enum": [
          0,
          1
        ],
        "x-apifox": {
          "enumDescriptions": {
            "0": "已删除",
            "1": "未删除"
          }
        },
        "x-apifox-folder": "enums"
      },
      "UpdateProductBrandReq": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "title": "品牌名称",
            "minLength": 1
          },
          "description": {
            "type": "string",
            "title": "品牌描述"
          },
          "picUrl": {
            "type": "string",
            "title": "品牌名图片"
          },
          "status": {
            "title": "状态",
            "description": "0->关闭;1->开启;",
            "$ref": "#/components/schemas/StatusEnum"
          }
        },
        "x-apifox-orders": [
          "name",
          "description",
          "picUrl",
          "status"
        ],
        "x-apifox-ignore-properties": [],
        "x-apifox-folder": "mall/product/brand"
      },
...
    }
  }
}